all_detais_table_updated

+++ src/com/spm/timetablemanagement/form/student_details/enterYnS.java
@@ -22,7 +22,8 @@ public class enterYnS extends javax.swing.JPanel {
 
     private Connection connection;
     private int yNsId;
-    
+    PreparedStatement pstYns;
+    ResultSet rs;
     /**
      * Creates new form enterYnS
      */
@@ -268,7 +269,7 @@ public class enterYnS extends javax.swing.JPanel {
     private void tbl_yNsMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tbl_yNsMouseClicked
         // TODO add your handling code here:
 
-        int i = tbl_yNs.getSelectedRow();
+       int i = tbl_yNs.getSelectedRow();
         TableModel model = tbl_yNs.getModel();
         txt_id.setText(model.getValueAt(i, 0).toString());
         txt_YnS.setText(model.getValueAt(i, 1).toString());
@@ -277,6 +278,8 @@ public class enterYnS extends javax.swing.JPanel {
     private void btn_deleteYnSActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_deleteYnSActionPerformed
         // TODO add your handling code here:
         String id = txt_id.getText();
+        int whileWorking = 0;
+               
 
         try
         {
@@ -288,13 +291,47 @@ public class enterYnS extends javax.swing.JPanel {
             }
             else{
                 txt_error.setText("");
-//            DefaultTableModel model = (DefaultTableModel)tbl_yNs.getModel();
-//            Statement smt = connection.createStatement();
-            smt.execute("DELETE FROM academic_year_and_semester WHERE id = "+id); 
+             
+            
+            String getYnsQuery = "select yNsId from generated_group_id WHERE yNsId = '"+id+"'";
+            pstYns = connection.prepareStatement(getYnsQuery);
+            rs = pstYns.executeQuery();
+            while(rs.next())
+            {    
+
+                    int x = JOptionPane.showConfirmDialog(this,"You have this related data,is it ok?", "Confirm", JOptionPane.YES_NO_OPTION);
+                    if (x == 0){
+                        ResultSet rs = smt.executeQuery("select id from generated_group_id where yNsId = "+id);
+                        String id_G="";
+                        while(rs.next()){
+                            id_G = rs.getString(1);
+                            System.out.println(id_G);
+                        }
+                        smt.execute("DELETE FROM generated_sub_group_id WHERE gId ="+Integer.parseInt(id_G));
+                        smt.execute("DELETE FROM generated_group_id WHERE yNsId = "+id); 
+                        smt.execute("DELETE FROM academic_year_and_semester WHERE id = "+id);
+                        smt.execute("DELETE FROM all_details WHERE yNs = '"+txt_YnS.getText().toString()+"'");
+                        
+                        JOptionPane.showMessageDialog(this, "Record Deleted!");
+
+                        
+                    }
+                    else{
+                        JOptionPane.showMessageDialog(this, "Delete Canceled!");
+                        
+                    }
+                    whileWorking++;
+                
+            }
+            if(whileWorking == 0){
+                smt.execute("DELETE FROM academic_year_and_semester WHERE id = "+id);
+                JOptionPane.showMessageDialog(this, "Record Deleted!");
+
+            }
             model.setRowCount(0);
             showYnSList();
             txt_YnS.setText("");
-            JOptionPane.showMessageDialog(this, "Record Deleted!");
+
             }
         }
         catch(Exception e)
